Use django-guardian for profile_detail view
Use the permission_required_or_404 decorator to restrict access to the
profile_detail view. There are actually checks inside the view to
restrict access to profiles. Accessing your own profile while not being
logged in does result in the string "You don't have permission to access
the profile" being displayed and the user is forced to press the back
button and login.
Adding the decorator to the view leaves the possibility to redirect the
user back to the login page by configuring django-guardian.